the amount data type is modified

+++ src/main/java/com/cubanfardo/cubanfardoreport/model/Shipping.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record Shipping(@NotBlank(message = "Tracking is mandatory") String tracking,
                        @NotBlank(message = "Province is mandatory") String province,
                        @PositiveOrZero(message = "Amount is positive value")
-                       @NotNull(message = "Amount is mandatory") Integer amount,
+                       @NotNull(message = "Amount is mandatory") Double amount,
                        Double weight, String details,
                        @NotNull(message = "Sender is mandatory") Sender sender,
                        @NotNull(message = "Receives is mandatory") Receives receives,
